+++ src/wink/tests/windows.cpp
@@ -10,7 +10,7 @@ int main( int argc, char *argv[] )
     try
     {
         wink::rand32_kiss g;
-        g.seed( time(NULL) );
+        g.seed( uint32_t(time(NULL)) );
         const double L = 10.0;
         const double Lo = -1;
         const double Hi = L+Lo;
